Move assert SourceData/KeyData functions to mockdata module
